create README
Create readme for better idea and scope of the project, leading to less consumer confusion when attempting to use the product.

This project is not explicitly clear about it's purpose and design goals despite it's fantasticly punny, yet juvenile name. "Readme" is extremely important for projects to be taken seriously and get widespread adopted use, especially if the project is hoped to be acquired by a larger company.